PSR-392 refactoring
