+++ app/controllers/nonsipp/membercontributions/MemberContributionsController.scala
@@ -14,15 +14,16 @@
  * limitations under the License.
  */
 
-package controllers.nonsipp.memberpayments
+package controllers.nonsipp.membercontributions
 
 import controllers.actions._
-import controllers.nonsipp.memberpayments.MemberContributionsController._
+import controllers.nonsipp.membercontributions.MemberContributionsController._
+import controllers.nonsipp.membercontributions.routes
 import forms.YesNoPageFormProvider
 import models.Mode
 import models.SchemeId.Srn
 import navigation.Navigator
-import pages.nonsipp.memberpayments.MemberContributionsPage
+import pages.nonsipp.membercontributions.MemberContributionsPage
 import play.api.data.Form
 import play.api.i18n.{I18nSupport, MessagesApi}
 import play.api.mvc.{Action, AnyContent, MessagesControllerComponents}
