+++ conf/app.routes
@@ -154,12 +154,6 @@ POST        /:srn/did-scheme-have-bank-account                                  
 GET         /:srn/change-did-scheme-have-bank-account                                   controllers.nonsipp.schemedesignatory.ActiveBankAccountController.onPageLoad(srn: Srn, mode: Mode = CheckMode)
 POST        /:srn/change-did-scheme-have-bank-account                                   controllers.nonsipp.schemedesignatory.ActiveBankAccountController.onSubmit(srn: Srn, mode: Mode = CheckMode)
 
-GET         /:srn/member-contributions                                                  controllers.nonsipp.memberpayments.MemberContributionsController.onPageLoad(srn: Srn, mode: Mode = NormalMode)
-POST        /:srn/member-contributions                                                  controllers.nonsipp.memberpayments.MemberContributionsController.onSubmit(srn: Srn, mode: Mode = NormalMode)
-GET         /:srn/change-member-contributions                                           controllers.nonsipp.memberpayments.MemberContributionsController.onPageLoad(srn: Srn, mode: Mode = CheckMode)
-POST        /:srn/change-member-contributions                                           controllers.nonsipp.memberpayments.MemberContributionsController.onSubmit(srn: Srn, mode: Mode = CheckMode)
-
-
 GET         /:srn/transfer-out                                                          controllers.nonsipp.membertransferout.SchemeTransferOutController.onPageLoad(srn: Srn, mode: Mode = NormalMode)
 POST        /:srn/transfer-out                                                          controllers.nonsipp.membertransferout.SchemeTransferOutController.onSubmit(srn: Srn, mode: Mode = NormalMode)
 GET         /:srn/change-transfer-out                                                   controllers.nonsipp.membertransferout.SchemeTransferOutController.onPageLoad(srn: Srn, mode: Mode = CheckMode)
