+++ conf/membercontributions.routes
@@ -1,3 +1,8 @@
+GET         /:srn/member-contributions                                              controllers.nonsipp.membercontributions.MemberContributionsController.onPageLoad(srn: Srn, mode: Mode = NormalMode)
+POST        /:srn/member-contributions                                              controllers.nonsipp.membercontributions.MemberContributionsController.onSubmit(srn: Srn, mode: Mode = NormalMode)
+GET         /:srn/change-member-contributions                                       controllers.nonsipp.membercontributions.MemberContributionsController.onPageLoad(srn: Srn, mode: Mode = CheckMode)
+POST        /:srn/change-member-contributions                                       controllers.nonsipp.membercontributions.MemberContributionsController.onSubmit(srn: Srn, mode: Mode = CheckMode)
+
 GET         /:srn/what-you-will-need-member-contributions                           controllers.nonsipp.membercontributions.WhatYouWillNeedMemberContributionsController.onPageLoad(srn: Srn)
 GET         /:srn/submit-what-you-will-need-member-contributions                    controllers.nonsipp.membercontributions.WhatYouWillNeedMemberContributionsController.onSubmit(srn: Srn)
